feat(#73):  history api, query 추가 + History 컴포넌트 components 폴더로 이동
